requirements.txt
playwright==1.44.0
gspread==6.1.2
google-auth==2.29.0
aiohttp==3.9.5
